Implement all neccessary routes
