+++ controllers/events.js
@@ -1,30 +1,8 @@
-import User from '../models/user.js'
-import Event from '../models/event.js'
+import User from '../models/user.js';
+import Event from '../models/event.js';
 
-// export async function getEventById(req, res) {
-//   try {
-//     const eventId = req.params.id;
-//     const event = await Event.findById(eventId);
-//     if (!event) {
-//       return res.status(404).json({ message: 'Event not found' });
-//     }
-//     res.status(200).json(event);
-//   } catch (error) {
-//     console.error(error);
-//     res.status(500).json({ message: 'Internal Server Error' });
-//   }
-// }
-
-// export async function getEvents(req, res) {
-//   try {
-//     const events = await Event.find();
-//     res.status(200).json(events);
-//   } catch (error) {
-//     console.error(error);
-//     res.status(500).json({ message: 'Internal Server Error' });
-//   }
-// }
-
+export async function getEvents(req, res) {
+}
 
 export async function createEvent(req, res) {
   try {
@@ -45,4 +23,7 @@ export async function createEvent(req, res) {
     console.error(error);
     res.status(500).json({ message: 'Internal Server Error' });
   }
+}
+
+export async function deleteEvent(req, res) {
 }
