+++ controllers/users.js
@@ -1,20 +1,4 @@
-import User from '../models/user.js'
-// import Event from '../models/event.js'
-
-// export async function getUserById(req, res) {
-//   try {
-//     const userId = req.params.id;
-//     const user = await User.findById(userId);
-
-//     if (!user) {
-//       return res.status(404).json({ message: 'User not found' });
-//     }
-//     res.status(200).json(user);
-//   } catch (error) {
-//     console.error(error);
-//     res.status(500).json({ message: 'Internal Server Error' });
-//   }
-// }
+import User from '../models/user.js';
 
 export async function createUser(req, res) {
   const userProfile = req.body;
