+++ models/event.js
@@ -33,24 +33,7 @@ const eventSchema = new mongoose.Schema({
     type: String,
     enum: ['Music', 'Sports', 'Food', 'Arts', 'Other'],
     required: true
-  },
-  likes: {
-    type: Number,
-    default: 0
-  },
-  dislikes: {
-    type: Number,
-    default: 0
-  },
-  //tracks which users have liked or disliked an event. ensure that they can only like an event once
-  usersLiked: [{
-    type: mongoose.Schema.Types.ObjectId,
-    ref: 'User'
-  }],
-  usersDisliked: [{
-    type: mongoose.Schema.Types.ObjectId,
-    ref: 'User'
-  }]
+  }
 })
 
 export default mongoose.model('Event', eventSchema);
